refactor: clean up recipe chat action
